refactored added clause, var, cnf, dnf classes added writers to them added fully working API added arguments parsing
