+++ expression_forms/nnf.py
@@ -1,0 +1,4 @@
+def get_nnf(expr):
+    new_expr = expr
+    new_expr.make_nnf()
+    return new_expr
