+++ main.py
@@ -1,16 +1,28 @@
-import expressions, logic
+import argparse
 
-cur_str = input()
-expr = expressions.read_expr(cur_str)
-expressions.print_left_to_right(expr)
-print()
-expr.make_nnf()
+from writers.nf_writer import print_nnf_from_string, print_cnf_from_string, print_dnf_from_string
 
-print("nnf")
-expressions.print_left_to_right(expr)
+parser = argparse.ArgumentParser()
+parser.add_argument("input", help="Input expression")
+parser.add_argument("--nnf", help="when checked outputs negative normal form",
+                    action="store_true")
+parser.add_argument("--cnf", help="when checked outputs conjunctive normal form",
+                    action="store_true")
+parser.add_argument("--dnf", help="when checked outputs disjunctive normal form",
+                    action="store_true")
+args = parser.parse_args()
 
+actions = [
+    ('nnf', args.nnf, print_nnf_from_string),
+    ('cnf', args.cnf, print_cnf_from_string),
+    ('dnf', args.dnf, print_dnf_from_string),
+]
 
-print("\ndnf")
-logic.print_dnf_str_appearance_from_expr(expr)
-print("\ncnf")
-logic.print_cnf_str_appearance_from_expr(expr)
+show_all = False
+if not (args.nnf or args.dnf or args.cnf):
+    show_all = True
+
+for action in actions:
+    if show_all or action[1]:
+        print(f"{action[0]}: ", end='')
+        action[2](args.input)
